improved codestyle and structure
